Fixed an LFR bug
Some LFR bosses showed as not being killed, when in fact they were. It
turns out that the encounter IDs are not ordered like I thought they
were, so I have to scan each raid portion for all encounters, and if any
of them are marked as killed, then record that.