Fix post-7.1 error due to API changes
The LFR API has changed, and instead of returning 7 encounters per wing
(with only a few being "real"), it now only returns the exact number of
bosses (with all being "real"). This means that I can now rely on the
actual numEncounters returned by the API to loop over per wing, and that
I need to use an offset for tracking boss id per overall raid. So,
Xavius is 7 in my list, even though he is the 1st (and only) encounter
in Rift of Aln.

Also noticed that my SavedVariables name was completely wrong, a
copy/paste error that's been there for quite some time. Perhaps it's not
actually necessary?